ml-service/requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
flask-cors==4.0.0
pycryptodome==3.19.0 